+++ extensions/git/src/components/views/StatusView.tsx
@@ -27,6 +27,7 @@ import { basename } from "path";
 import { BranchAttachedLinksAction, BranchPushAction, BranchPushForceAction } from "../actions/BranchActions";
 import { RemoteFetchAction, RemotePullAction } from "../actions/RemoteActions";
 import { CopyToClipboardMenuAction } from "../actions/CopyToClipboardMenuAction";
+import { GitIgnoreAction } from "../actions/GitIgnoreAction";
 
 export function StatusView(context: RepositoryContext & NavigationContext) {
   const toggleController = useToggleDetail("Status Diff", "Changes", true);
@@ -80,6 +81,7 @@ export function StatusView(context: RepositoryContext & NavigationContext) {
               </ActionPanel.Section>
 
               <ToggleDetailAction controller={toggleController} />
+              <GitIgnoreAction {...context} />
               <WorkspaceNavigationActions {...context} />
             </ActionPanel>
           }
@@ -115,6 +117,7 @@ export function StatusView(context: RepositoryContext & NavigationContext) {
                 {context.branches.data.currentBranch && (
                   <BranchAttachedLinksAction {...context} branch={context.branches.data.currentBranch} />
                 )}
+                <GitIgnoreAction {...context} />
               </ActionPanel.Section>
 
               <WorkspaceNavigationActions {...context} />
@@ -278,6 +281,7 @@ function FileListItem(
               </>
             )}
             <FileHistoryAction filePath={context.file.absolutePath} {...context} />
+            <GitIgnoreAction filePath={context.file.relativePath} {...context} />
           </ActionPanel.Section>
 
           <ActionPanel.Section>